requirements.txt
# AURORA-XT Vietnamese AVSR - Requirements
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0

transformers>=4.36.0
timm>=0.9.12
huggingface-hub>=0.20.0

# decord>=0.6.0
av>=11.0.0
soundfile>=0.12.1
opencv-python-headless>=4.9.0

mediapipe>=0.10.9


webdataset>=0.2.86
numpy<2
Pillow>=10.2.0

jiwer>=3.0.3
unicodedata2>=15.0.0

tqdm>=4.66.0
wandb>=0.16.0
pyyaml>=6.0.1

modal>=0.56.0

pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
